using threads

+++ thumbnail_maker.py
@@ -4,39 +4,58 @@ import os
 import logging
 from urllib.parse import urlparse
 from urllib.request import urlretrieve
+from queue import Queue
+from threading import Thread
+
 
 import PIL
 from PIL import Image
 
-logging.basicConfig(filename='logfile.log', level=logging.DEBUG)
+FORMAT = "[%(threadName)s, %(asctime)s, %(levelname)s] %(message)s"
+logging.basicConfig(filename='logfile.log', level=logging.DEBUG, format=FORMAT)
 
 class ThumbnailMakerService(object):
     def __init__(self, home_dir='.'):
         self.home_dir = home_dir
         self.input_dir = self.home_dir + os.path.sep + 'incoming'
         self.output_dir = self.home_dir + os.path.sep + 'outgoing'
+        self.dl_queue = Queue()
+        self.img_queue = Queue()
 
-    def download_images(self, img_url_list):
-        # validate inputs
-        if not img_url_list:
-            return
-        os.makedirs(self.input_dir, exist_ok=True)
+    def download_image(self):
+        while not self.dl_queue.empty():
+            try:
+                url = self.dl_queue.get(block=False)
+                img_filename = urlparse(url).path.split('/')[-1]
+                urlretrieve(url, self.input_dir + os.path.sep + img_filename)
+                self.img_queue.put(img_filename)
+
+                self.dl_queue.task_done()
+            except Queue.Empty:
+                logging.info("Queue empty")
+
+    # def download_images(self, img_url_list):
+    #     # validate inputs
+    #     if not img_url_list:
+    #         return
+    #     os.makedirs(self.input_dir, exist_ok=True)
         
-        logging.info("beginning image downloads")
+    #     logging.info("beginning image downloads")
 
-        start = time.perf_counter()
-        for url in img_url_list:
-            # download each image and save to the input dir 
-            img_filename = urlparse(url).path.split('/')[-1]
-            urlretrieve(url, self.input_dir + os.path.sep + img_filename)
-        end = time.perf_counter()
+    #     start = time.perf_counter()
+    #     for url in img_url_list:
+    #         # download each image and save to the input dir 
+    #         img_filename = urlparse(url).path.split('/')[-1]
+    #         urlretrieve(url, self.input_dir + os.path.sep + img_filename)
+    #         self.img_queue.put(img_filename)
+    #     end = time.perf_counter()
 
-        logging.info("downloaded {} images in {} seconds".format(len(img_url_list), end - start))
+    #     self.img_queue.put(None)
+
+    #     logging.info("downloaded {} images in {} seconds".format(len(img_url_list), end - start))
 
     def perform_resizing(self):
         # validate inputs
-        if not os.listdir(self.input_dir):
-            return
         os.makedirs(self.output_dir, exist_ok=True)
 
         logging.info("beginning image resizing")
@@ -44,22 +63,31 @@ class ThumbnailMakerService(object):
         num_images = len(os.listdir(self.input_dir))
 
         start = time.perf_counter()
-        for filename in os.listdir(self.input_dir):
-            orig_img = Image.open(self.input_dir + os.path.sep + filename)
-            for basewidth in target_sizes:
-                img = orig_img
-                # calculate target height of the resized image to maintain the aspect ratio
-                wpercent = (basewidth / float(img.size[0]))
-                hsize = int((float(img.size[1]) * float(wpercent)))
-                # perform resizing
-                img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
-                
-                # save the resized image to the output dir with a modified file name 
-                new_filename = os.path.splitext(filename)[0] + \
-                    '_' + str(basewidth) + os.path.splitext(filename)[1]
-                img.save(self.output_dir + os.path.sep + new_filename)
+        while True:
+            filename = self.img_queue.get()
+            if filename:
+                logging.info("resizing image {}".format(filename))
+                orig_img = Image.open(self.input_dir + os.path.sep + filename)
+                for basewidth in target_sizes:
+                    img = orig_img
+                    # calculate target height of the resized image to maintain the aspect ratio
+                    wpercent = (basewidth / float(img.size[0]))
+                    hsize = int((float(img.size[1]) * float(wpercent)))
+                    # perform resizing
+                    img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
+                    
+                    # save the resized image to the output dir with a modified file name 
+                    new_filename = os.path.splitext(filename)[0] + \
+                        '_' + str(basewidth) + os.path.splitext(filename)[1]
+                    img.save(self.output_dir + os.path.sep + new_filename)
 
-            os.remove(self.input_dir + os.path.sep + filename)
+                os.remove(self.input_dir + os.path.sep + filename)
+                logging.info("done resizing")
+                self.img_queue.task_done()
+            else:
+                self.img_queue.task_done()
+                break
+
         end = time.perf_counter()
 
         logging.info("created {} thumbnails in {} seconds".format(num_images, end - start))
@@ -68,8 +96,20 @@ class ThumbnailMakerService(object):
         logging.info("START make_thumbnails")
         start = time.perf_counter()
 
-        self.download_images(img_url_list)
-        self.perform_resizing()
+        for img_url in img_url_list:
+            self.dl_queue.put(img_url)
+
+        num_dl_threads = 4
+        for _ in range(num_dl_threads):
+            t = Thread(target=self.download_image)
+            t.start()
+        
+        t2 = Thread(target=self.perform_resizing)
+        t2.start()
+
+        self.dl_queue.join()
+        self.img_queue.put(None)
+        t2.join()
 
         end = time.perf_counter()
         logging.info("END make_thumbnails in {} seconds".format(end - start))
